Rename _ANN_&_CNN to _ANN_+_CNN
#1 Artificial Neural Networks or (ANNs): 

Artificial Neural Networks or (ANNs) as we define them are a cluster of pipelines and connections which are made to LEARN a particular set of actions without the need for any continuous programming or set of codes for each different iteration. This makes them quite similar in structure and performance to the neurons we have in our brain. These neurons are those body parts in our brain which take the visual data as input from what our eyes see, interpret it and then send electrical impulses in our brain to perform a task that helps to objectify or recognize the pattern we are currently observing. An ANN has based on similar structures called nodes which are units.


#2 Convolutional Neural Network (CNN): 


Within the field of deep learning, a CNN or ConvNet as it is commonly referred to is a class of deep non-cyclic artificial neural network which find its extensive use in analyzing visual imagery.

It uses multiple phases of multilayer perceptrons and is built on the basis to work with minimal pre-processing. The CNN is different from other versions of an artificial neural network as it utilizes very few amounts of pre-processing relative to other image classification algorithms. It holds an advantage in the fact that it automates the filters itself which in previous versions were hand-built.

So how does this help me? CNN architecture runs on a pre-assumed task that it recognizes the input as images, that in later phases is beneficial to us as we can set up some characters and features into the architecture that required another set of datasets earlier. Thus, the non-linearity of the functions become more efficient and the values to be provided are computed by themselves.